+++ semcore/d3-chart/src/component/Chart/DonutChart.type.ts
@@ -5,7 +5,9 @@ import type React from 'react';
 
 import type { BaseChartProps } from './AbstractChart.type';
 
-export type DonutChartData = Record<string, number>;
+export type DataKey = string;
+
+export type DonutChartData = Record<DataKey, number>;
 
 export type DonutChartProps = BaseChartProps<DonutChartData> & {
   /** Internal */
@@ -20,6 +22,8 @@ export type DonutChartProps = BaseChartProps<DonutChartData> & {
   halfsize?: boolean;
   /** Content displayed in the center of the donut */
   innerLabel?: React.ReactNode;
+  /** Callback triggered when a user clicks on a pie */
+  onClickPie?: (key: DataKey, e: React.SyntheticEvent) => void;
 };
 
 export type DonutChartType = Intergalactic.Component<typeof Flex, DonutChartProps>;