+++ semcore/d3-chart/src/component/Chart/LineChart.type.ts
@@ -30,6 +30,8 @@ export type LineChartProps = BaseChartProps<LineChartData> & {
   curve?: CurveFactory;
   /** Curve factory specifically for area rendering */
   areaCurve?: CurveFactory;
+  /** Callback triggered when a user clicks on a line */
+  onClickLine?: (index: number, event: React.SyntheticEvent) => void;
 };
 
 export type LineChartType = Intergalactic.Component<typeof Flex, LineChartProps>;