+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -27,7 +27,7 @@ import { localizedMessages } from '../../translations/__intergalactic-dynamic-lo
 import { Body } from '../Body/Body';
 import type { BodyPropsInner } from '../Body/Body.types';
 import { MergedColumnsCell, MergedRowsCell } from '../Body/MergedCells';
-import type { DTRow, UniqRowKey } from '../Body/Row.types';
+import type { DTRow } from '../Body/Row.types';
 import type { DataTableColumnProps, DTColumn } from '../Head/Column.types';
 import type { DataTableGroupProps } from '../Head/Group.type';
 import { Head } from '../Head/Head';
@@ -47,13 +47,17 @@ type State = {
   selectAllMessage: string;
 };
 
-class DataTableRoot<D extends DataTableData> extends Component<
-  DataTableProps<D>,
-  {},
-  {},
+class DataTableRoot<
+  Data extends DataTableData,
+  UniqKey extends keyof Data[number],
+  UniqKeyType extends Data[number][UniqKey],
+> extends Component<
+    DataTableProps<Data, UniqKey, UniqKeyType>,
+    {},
+    {},
   typeof DataTableRoot.enhance,
-  { use: DTRow; expandedRows: Set<string>; renderEmptyData: () => React.ReactNode }
-> {
+  { use: DTRow<UniqKeyType>; expandedRows: Set<UniqKeyType>; renderEmptyData: () => React.ReactNode }
+  > {
   static displayName = 'DataTable';
   static style = style;
 
@@ -84,14 +88,14 @@ class DataTableRoot<D extends DataTableData> extends Component<
   private gridAreaGroupMap = new Map<number, string>();
 
   private columnsSplitter = '/';
-  private rows: Array<DTRow | DTRow[]> = [];
-  private flatRows: DTRow[] = [];
+  private rows: Array<DTRow<UniqKeyType> | DTRow<UniqKeyType>[]> = [];
+  private flatRows: DTRow<UniqKeyType>[] = [];
 
   private selectAllMessageTimer = 0;
 
   private headerNodesMap = new Map();
 
-  constructor(props: DataTableProps<D>) {
+  constructor(props: DataTableProps<Data, UniqKey, UniqKeyType>) {
     super(props);
 
     if (props.children) {
@@ -119,7 +123,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
   }
 
   componentDidMount() {
-    const { headerProps, loading } = this.asProps;
+    const { headerProps, loading, selectedRows } = this.asProps;
     if ((headerProps?.sticky && !headerProps.h) || loading || this.columns.some((c) => c.fixed)) {
       requestAnimationFrame(() => {
         this.forceUpdate();
@@ -197,7 +201,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
     };
   }
 
-  getHeadProps(): HeadPropsInner<D> {
+  getHeadProps(): HeadPropsInner<Data, UniqKey, UniqKeyType> {
     const {
       use,
       compact,
@@ -228,14 +232,19 @@ class DataTableRoot<D extends DataTableData> extends Component<
       gridTemplateAreas,
       sideIndents,
       totalRows: this.totalRows,
-      selectedRows: selectedRows,
+      selectedRows,
+      flatRows: this.flatRows,
       onChangeSelectAll: (value, e) => {
-        if (value === false) {
-          onSelectedRowsChange?.([], e);
+        const mappedFlatRows = this.flatRows.map((r) => r[UNIQ_ROW_KEY]);
+        const selectedRowsSet = new Set(selectedRows);
+
+        if (value) {
+          mappedFlatRows.forEach(selectedRowsSet.add, selectedRowsSet);
         } else {
-          const selectedRows = this.flatRows.map((row) => row[UNIQ_ROW_KEY]);
-          onSelectedRowsChange?.(selectedRows, e);
+          mappedFlatRows.forEach(selectedRowsSet.delete, selectedRowsSet);
         }
+
+        onSelectedRowsChange?.(Array.from(selectedRowsSet), e);
       },
       getFixedStyle: this.getFixedStyle,
       onCellClick: this.handleCellClick,
@@ -243,7 +252,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
     };
   }
 
-  getBodyProps(): BodyPropsInner {
+  getBodyProps(): BodyPropsInner<UniqKeyType> {
     const {
       use,
       compact,
@@ -259,6 +268,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
       sideIndents,
       selectedRows,
       accordionDuration,
+      data: rawData,
     } = this.asProps;
     const { gridTemplateColumns, gridTemplateAreas } = this.gridSettings;
     return {
@@ -294,12 +304,12 @@ class DataTableRoot<D extends DataTableData> extends Component<
       onSelectRow: this.handleSelectRow,
       getFixedStyle: this.getFixedStyle,
       onCellClick: this.handleCellClick,
+      rawData,
     };
   }
 
-  handleCellClick = (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow }) => {
+  handleCellClick = (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow<UniqKeyType> }) => {
     if (lastInteraction.isMouse()) {
-      console.log('click on cell in data table handler', opt);
       this.initFocusableCell([this.hasFocusableInHeader() ? opt.rowIndex + 1 : opt.rowIndex, opt.colIndex]);
     }
   };
@@ -307,26 +317,22 @@ class DataTableRoot<D extends DataTableData> extends Component<
   handleSelectRow = (
     isSelected: boolean,
     selectedRowIndex: number,
-    row: DTRow,
+    row: DTRow<UniqKeyType>,
     event?: React.SyntheticEvent<HTMLElement>,
   ) => {
     const { selectedRows, onSelectedRowsChange } = this.asProps;
 
-    if (selectedRows && onSelectedRowsChange) {
-      const newSelectedRows = new Set(selectedRows);
+    if (!selectedRows || !onSelectedRowsChange) return;
 
-      if (isSelected && !newSelectedRows.has(row[UNIQ_ROW_KEY])) {
-        newSelectedRows.add(row[UNIQ_ROW_KEY]);
-      } else if (!isSelected && newSelectedRows.has(row[UNIQ_ROW_KEY])) {
-        newSelectedRows.delete(row[UNIQ_ROW_KEY]);
-      }
+    const selectedRowsSet = new Set(selectedRows);
 
-      onSelectedRowsChange([...newSelectedRows], event, {
-        selectedRowIndex,
-        isSelected,
-        row,
-      });
+    if (selectedRowsSet.has(row[UNIQ_ROW_KEY])) {
+      selectedRowsSet.delete(row[UNIQ_ROW_KEY]);
+    } else {
+      selectedRowsSet.add(row[UNIQ_ROW_KEY]);
     }
+
+    onSelectedRowsChange(Array.from(selectedRowsSet), event, { selectedRowIndex, isSelected, row });
   };
 
   setSelectAllMessage = (selectedAll: boolean) => {
@@ -372,7 +378,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
       this.columns.some((column) => column.sortable);
   };
 
-  onExpandRow = (expandedRow: DTRow) => {
+  onExpandRow = (expandedRow: DTRow<UniqKeyType>) => {
     const { expandedRows } = this.asProps;
     if (expandedRows.has(expandedRow[UNIQ_ROW_KEY])) {
       expandedRows.delete(expandedRow[UNIQ_ROW_KEY]);
@@ -964,7 +970,11 @@ class DataTableRoot<D extends DataTableData> extends Component<
   private calculateColumnsFromConfig(): [DTColumn[], DTColumn[]] {
     const { columns, data, selectedRows } = this.props;
 
-    this.hasGroups = columns.some((column) => 'columns' in column);
+    this.hasGroups = columns.some((column) => {
+      return 'columns' in column && column.columns.some((col) => {
+        return col.children !== null;
+      });
+    });
 
     let groupIndex = 0;
     let gridColumnIndex = selectedRows ? 2 : 1;
@@ -990,6 +1000,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
       parent?: DTColumn,
       isFirst?: boolean,
       isLast?: boolean,
+      hasGroups?: boolean,
     ): DTColumn => {
       const leftBordersFromParent =
         isFirst && (parent?.borders === 'both' || parent?.borders === 'left') ? 'left' : undefined;
@@ -1001,7 +1012,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
 
         name: childIsColumn(columnElement) ? columnElement.name : '',
         gtcWidth: childIsColumn(columnElement) ? calculateGridTemplateColumn(columnElement) : '',
-        fixed: columnElement.fixed ?? parent?.fixed,
+        fixed: columnElement.fixed ?? (hasGroups ? parent?.fixed : undefined),
         borders: columnElement.borders ?? leftBordersFromParent ?? rightBordersFromParent,
         parent,
       } as DTColumn;
@@ -1038,21 +1049,23 @@ class DataTableRoot<D extends DataTableData> extends Component<
 
         const initGridColumn = gridColumnIndex;
 
+        const groupedRow = this.hasGroups ? 2 : 1;
+
         Group.columns = [];
         Group.children = child.children;
         child.columns.forEach((child, j) => {
           const isFirst = j === 0;
           const isLast = j === childCount - 1;
-          const col = makeColumn(child, Group, isFirst, isLast);
+          const col = makeColumn(child, Group, isFirst, isLast, this.hasGroups);
 
           if (i === 0 && j === 0 && data.some((d) => d[ACCORDION])) {
             gridColumnIndex++;
-            col.gridArea = `2 / ${gridColumnIndex - 1} / 3 / ${gridColumnIndex + 1}`;
+            col.gridArea = `${groupedRow} / ${gridColumnIndex - 1} / ${groupedRow + 1} / ${gridColumnIndex + 1}`;
           } else {
-            col.gridArea = `2 / ${gridColumnIndex} / 3 / ${gridColumnIndex + 1}`;
+            col.gridArea = `${groupedRow} / ${gridColumnIndex} / ${groupedRow + 1} / ${gridColumnIndex + 1}`;
           }
 
-          col.gridArea = `2 / ${gridColumnIndex} / 3 / ${gridColumnIndex + 1}`;
+          col.gridArea = `${groupedRow} / ${gridColumnIndex} / ${groupedRow + 1} / ${gridColumnIndex + 1}`;
           gridColumnIndex++;
 
           calculatedColumns.push(col);
@@ -1070,12 +1083,12 @@ class DataTableRoot<D extends DataTableData> extends Component<
     return [calculatedColumns, treeColumns];
   }
 
-  private calculateRows(): Array<DTRow[] | DTRow> {
+  private calculateRows(): Array<DTRow<UniqKeyType>[] | DTRow<UniqKeyType>> {
     const columns = this.columns;
     // @ts-ignore
     const { data, uid, uniqueRowKey } = this.props;
 
-    const rows: Array<DTRow[] | DTRow> = [];
+    const rows: Array<DTRow<UniqKeyType>[] | DTRow<UniqKeyType>> = [];
     const columnNames = columns.map((column: DTColumn) => column.name);
 
     let rowIndex = 0;
@@ -1084,7 +1097,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
 
     const makeDtRow = (row: DataRowItem, excludeColumns?: string[]) => {
       const columns = new Set(columnNames);
-      const dtRow = Object.entries(row).reduce<DTRow>(
+      const dtRow = Object.entries(row).reduce<DTRow<UniqKeyType>>(
         (acc, [key, value]) => {
           const columnsToRow = key.split(this.columnsSplitter);
 
@@ -1108,8 +1121,12 @@ class DataTableRoot<D extends DataTableData> extends Component<
           return acc;
         },
         {
-          [UNIQ_ROW_KEY]:
-            row[UNIQ_ROW_KEY] || (uniqueRowKey ? row[uniqueRowKey] : `${uid}_${(rowIndex + id).toString(36)}`) as UniqRowKey,
+          /*
+            row -> DataRowItem
+            uniqueRowKey is a `keyof Data[number]` -> `keyof DataRowItem`
+          */
+          // @ts-ignore
+          [UNIQ_ROW_KEY]: row[UNIQ_ROW_KEY] || (uniqueRowKey ? row[uniqueRowKey] : `${uid}_${(rowIndex + id).toString(36)}`),
           [ROW_INDEX]: rowIndex,
         },
       );
@@ -1131,10 +1148,10 @@ class DataTableRoot<D extends DataTableData> extends Component<
       const groupedRows: DataTableData | undefined = row[ROW_GROUP];
 
       if (groupedRows) {
-        const innerRows: DTRow[] & { [ACCORDION]?: React.ReactElement } = [];
+        const innerRows: DTRow<UniqKeyType>[] & { [ACCORDION]?: React.ReactElement } = [];
 
         const groupedKeys: string[] = [];
-        const groupedRowData = Object.entries(row).reduce<Omit<DTRow, symbol>>(
+        const groupedRowData = Object.entries(row).reduce<Omit<DTRow<UniqKeyType>, symbol>>(
           (acc, [key, value]) => {
             acc[key] = new MergedRowsCell(value, groupedRows.length, row[ACCORDION]);
             groupedKeys.push(key);
@@ -1147,7 +1164,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
         );
 
         groupedRows.forEach((childRow, index) => {
-          let dtRow: DTRow;
+          let dtRow: DTRow<UniqKeyType>;
           if (index === 0) {
             const rowData = {
               ...childRow,