+++ semcore/data-table/src/components/Head/Head.types.ts
@@ -1,6 +1,6 @@
 import type { DTColumn } from './Column.types';
 import type { CellPropsInner } from '../Body/Cell.types';
-import type { UniqRowKey } from '../Body/Row.types';
+import type { DTRow } from '../Body/Row.types';
 import type { DataTableData, DataTableProps, DTUse } from '../DataTable/DataTable.types';
 
 export type DataTableHeadProps = {
@@ -22,16 +22,23 @@ export type DataTableHeadProps = {
 
   /** Enable scroll bar element in header */
   withScrollBar?: boolean;
+
+  /** Animation duration for a top property. When it's set adds animation for top changes */
+  animationDuration?: number;
 };
 
-export type HeadPropsInner<D extends DataTableData> = {
+export type HeadPropsInner<
+  Data extends DataTableData,
+  UniqKey extends keyof Data[number],
+  UniqKeyType extends Data[number][UniqKey],
+> = {
   use: DTUse;
   tableRef: React.RefObject<HTMLElement>;
   columns: DTColumn[];
   treeColumns: DTColumn[];
   compact: boolean;
-  sort?: DataTableProps<D>['sort'];
-  onSortChange?: DataTableProps<D>['onSortChange'];
+  sort?: DataTableProps<Data, UniqKey, UniqKeyType>['sort'];
+  onSortChange?: DataTableProps<Data, UniqKey, UniqKeyType>['onSortChange'];
   getI18nText: (key: string) => string;
   uid: string;
   ref: React.RefObject<HTMLDivElement>;
@@ -42,11 +49,12 @@ export type HeadPropsInner<D extends DataTableData> = {
   sideIndents?: 'wide';
 
   totalRows: number;
-  selectedRows?: UniqRowKey[];
+  selectedRows?: UniqKeyType[];
   onChangeSelectAll?: (value: boolean, event?: React.SyntheticEvent<HTMLElement>) => void;
+  flatRows: DTRow<UniqKeyType>[];
 
   getFixedStyle: (
     cell: Pick<DTColumn, 'name' | 'fixed'>,
   ) => [side: 'left' | 'right', style: string | number] | [side: undefined, style: undefined];
-  onCellClick: CellPropsInner['onClick'];
+  onCellClick: CellPropsInner<UniqKeyType>['onClick'];
 };