+++ stories/components/d3-chart/tests/bubble-chart.stories.tsx
@@ -1,6 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import BubbleChartExample from './examples/bubble-chart/bubble-chart';
+import BubbleChartOnClickExample from './examples/bubble-chart/bubble-chart-on-click';
 
 const meta: Meta = {
   title: 'Components/d3Charts/Tests/Bubble-Chart',
@@ -10,4 +11,8 @@ export default meta;
 
 export const BubbleChart: StoryObj = {
   render: BubbleChartExample,
+};
+
+export const BubbleChartOnClick: StoryObj = {
+  render: BubbleChartOnClickExample,
 };