+++ stories/components/data-table/docs/examples/sorting-changing-size.tsx
@@ -25,7 +25,7 @@ const Demo = () => {
   );
 
   return (
-    <DataTable<typeof data>
+    <DataTable
       data={sortedData}
       sort={sort}
       onSortChange={setSort}
@@ -40,11 +40,12 @@ const Demo = () => {
           style: { whiteSpace: 'nowrap' },
           borders: 'both',
           sortable: true,
+          changeSortSize: true,
         },
         {
           name: 'cpc',
           children: 'CPC',
-          gtcWidth: 'minmax(0, 60px)',
+          gtcWidth: 'minmax(60px, 66px)',
           borders: 'right',
           sortable: true,
           changeSortSize: true,